Organise frontend folder structure
