+++ frontend/src/pages/user/Login.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginBox from "../components/LoginBox";
+import LoginBox from "../../components/LoginBox";
 
 const Login = ({ onLogin }) => {
   const handleLogin = (isAdmin) => {
